requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-otp==1.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
